requirements.txt
django-ipware
flake8
isort
